tasks controller change

+++ server/controllers/tasksController.js
@@ -2,6 +2,7 @@ const Task = require("./../models/taskModel");
 const catchAsync = require("./../utils/catchAsync");
 const APIFeatures = require("./../utils/apiFeatures");
 const AppError = require("./../utils/appError");
+const User = require("../models/userModel");
 
 exports.getAllTasks = catchAsync(async (req, res, next) => {
   const features = new APIFeatures(Task.find(), req.query).filter().sort();
@@ -41,37 +42,48 @@ exports.getTasks = catchAsync(async (req, res, next) => {
   });
 });
 
+//when updating a task, you are updating the task itself
 exports.updateTask = catchAsync(async (req, res, next) => {
-  const task = await Task.findByIdAndUpdate(req.params.id, req.body, {
+  const newTask = await Task.findByIdAndUpdate(req.params.id, req.body, {
     new: true,
     runValidators: true
   });
-  if (!task) {
+  if (!newTask) {
     return next(new AppError("No task with that id ", 404));
   }
   res.status(200).json({
     status: "success",
     data: {
-      task
+      newTask
     }
   });
 });
 
+//when deleting a task, you also need to delete the task from the tasks array of the user
 exports.deleteTask = catchAsync(async (req, res, next) => {
+  const arr = req.user.tasks;
+  const result = arr.find((el) => parseInt(el) === parseInt(req.params.id));
+  if (req.user.role === "user" && !result) {
+    return next(new AppError("You are not allowed to delete this task", 404));
+  }
   const task = await Task.findByIdAndDelete(req.params.id);
-
+  await User.updateOne(
+    { _id: req.user._id },
+    { $pull: { tasks: req.params.id } }
+  );
   if (!task) {
     return next(new AppError("No task found with that ID", 404));
   }
-
   res.status(204).json({
     status: "success",
-    data: null
+    data: "xyz"
   });
 });
 
+//when creating, the new id should be pushed to the tasks array- can be done using tasks.push(newTask) and doc.save()
 exports.createTask = catchAsync(async (req, res, next) => {
   const newTask = await Task.create(req.body);
+  await User.updateOne({ _id: req.user._id }, { $push: { tasks: newTask } });
   res.status(201).json({
     status: "success",
     data: {
